+++ src/main/java/sshj/sshj/service/ExpoPushService.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface ExpoPushService {
-    void sendingPushMeetingCreated(int clubId) throws Exception;
-    void sendingPushClubNoticeCreated(int clubId) throws Exception;
+    void sendingPushMeetingCreated(long clubId) throws Exception;
+    void sendingPushClubNoticeCreated(long clubId) throws Exception;
 }
